Airflow_Docker/requirements.txt
meteomatics
python-dotenv
psycopg2-binary
apache-airflow
datetime
pandas
sqlalchemy
requests